Refactor: arrange file
